pyproject.toml
[project]
name = "powerflow"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "pandas-stubs~=2.3.3",
]
